+++ src/network.rs
@@ -11,6 +11,7 @@ use rspotify::{
   client::Spotify,
   model::{
     album::SimplifiedAlbum,
+    artist::FullArtist,
     offset::for_position,
     page::Page,
     playlist::{PlaylistTrack, SimplifiedPlaylist},
@@ -74,9 +75,11 @@ pub enum IoEvent {
   GetRecommendationsForTrackId(String, Option<Country>),
   GetRecentlyPlayed,
   GetFollowedArtists(Option<String>),
+  SetArtistsToTable(Vec<FullArtist>),
   UserArtistFollowCheck(Vec<String>),
   GetAlbum(String),
-  SetDeviceIdInConfig(String),
+  TransferPlaybackToDevice(String),
+  GetAlbumForTrack(String),
   CurrentUserSavedTracksContains(Vec<String>),
   GetShowEpisodes(Option<String>, u32),
   AddItemToQueue(String),
@@ -254,14 +257,20 @@ impl<'a> Network<'a> {
       IoEvent::GetFollowedArtists(after) => {
         self.get_followed_artists(after).await;
       }
+      IoEvent::SetArtistsToTable(full_artists) => {
+        self.set_artists_to_table(full_artists).await;
+      }
       IoEvent::UserArtistFollowCheck(artist_ids) => {
         self.user_artist_check_follow(artist_ids).await;
       }
       IoEvent::GetAlbum(album_id) => {
         self.get_album(album_id).await;
       }
-      IoEvent::SetDeviceIdInConfig(device_id) => {
-        self.set_device_id_in_config(device_id).await;
+      IoEvent::TransferPlaybackToDevice(device_id) => {
+        self.transfert_playback_to_device(device_id).await;
+      }
+      IoEvent::GetAlbumForTrack(track_id) => {
+        self.get_album_for_track(track_id).await;
       }
       IoEvent::Shuffle(shuffle_state) => {
         self.shuffle(shuffle_state).await;
@@ -380,9 +389,7 @@ impl<'a> Network<'a> {
 
       let mut app = self.app.lock().await;
       app.playlist_tracks = Some(playlist_tracks);
-      if app.get_current_route().id != RouteId::TrackTable {
-        app.push_navigation_stack(RouteId::TrackTable, ActiveBlock::TrackTable);
-      };
+      app.push_navigation_stack(RouteId::TrackTable, ActiveBlock::TrackTable);
     };
   }
 
@@ -410,6 +417,11 @@ impl<'a> Network<'a> {
         .filter_map(|item| item.id)
         .collect::<Vec<String>>(),
     ));
+  }
+
+  async fn set_artists_to_table(&mut self, artists: Vec<FullArtist>) {
+    let mut app = self.app.lock().await;
+    app.artists = artists;
   }
 
   async fn get_made_for_you_playlist_tracks(
@@ -805,6 +817,15 @@ impl<'a> Network<'a> {
     if let Ok((albums, top_tracks, related_artist)) = try_join!(albums, top_tracks, related_artist)
     {
       let mut app = self.app.lock().await;
+
+      app.dispatch(IoEvent::CurrentUserSavedAlbumsContains(
+        albums
+          .items
+          .iter()
+          .filter_map(|item| item.id.to_owned())
+          .collect(),
+      ));
+
       app.artist = Some(Artist {
         artist_name,
         albums,
@@ -1001,10 +1022,10 @@ impl<'a> Network<'a> {
   }
 
   async fn user_artist_check_follow(&mut self, artist_ids: Vec<String>) {
-    if let Ok(are_follwed) = self.spotify.user_artist_check_follow(&artist_ids).await {
+    if let Ok(are_followed) = self.spotify.user_artist_check_follow(&artist_ids).await {
       let mut app = self.app.lock().await;
       artist_ids.iter().enumerate().for_each(|(i, id)| {
-        if are_follwed[i] {
+        if are_followed[i] {
           app.followed_artist_ids_set.insert(id.to_owned());
         } else {
           app.followed_artist_ids_set.remove(id);
@@ -1037,14 +1058,14 @@ impl<'a> Network<'a> {
   }
 
   async fn current_user_saved_albums_contains(&mut self, album_ids: Vec<String>) {
-    if let Ok(are_follwed) = self
+    if let Ok(are_followed) = self
       .spotify
       .current_user_saved_albums_contains(&album_ids)
       .await
     {
       let mut app = self.app.lock().await;
       album_ids.iter().enumerate().for_each(|(i, id)| {
-        if are_follwed[i] {
+        if are_followed[i] {
           app.saved_album_ids_set.insert(id.to_owned());
         } else {
           app.saved_album_ids_set.remove(id);
@@ -1272,7 +1293,50 @@ impl<'a> Network<'a> {
     }
   }
 
-  async fn set_device_id_in_config(&mut self, device_id: String) {
+  async fn get_album_for_track(&mut self, track_id: String) {
+    match self.spotify.track(&track_id).await {
+      Ok(track) => {
+        // It is unclear when the id can ever be None, but perhaps a track can be album-less. If
+        // so, there isn't much to do here anyways, since we're looking for the parent album.
+        let album_id = match track.album.id {
+          Some(id) => id,
+          None => return,
+        };
+
+        if let Ok(album) = self.spotify.album(&album_id).await {
+          // The way we map to the UI is zero-indexed, but Spotify is 1-indexed.
+          let zero_indexed_track_number = track.track_number - 1;
+          let selected_album = SelectedFullAlbum {
+            album,
+            // Overflow should be essentially impossible here, so we prefer the cleaner 'as'.
+            selected_index: zero_indexed_track_number as usize,
+          };
+
+          let mut app = self.app.lock().await;
+
+          app.selected_album_full = Some(selected_album.clone());
+          app.saved_album_tracks_index = selected_album.selected_index;
+          app.album_table_context = AlbumTableContext::Full;
+          app.push_navigation_stack(RouteId::AlbumTracks, ActiveBlock::AlbumTracks);
+        }
+      }
+      Err(e) => {
+        self.handle_error(anyhow!(e)).await;
+      }
+    }
+  }
+
+  async fn transfert_playback_to_device(&mut self, device_id: String) {
+    match self.spotify.transfer_playback(&device_id, true).await {
+      Ok(()) => {
+        self.get_current_playback().await;
+      }
+      Err(e) => {
+        self.handle_error(anyhow!(e)).await;
+        return;
+      }
+    };
+
     match self.client_config.set_device_id(device_id) {
       Ok(()) => {
         let mut app = self.app.lock().await;